Bug#37318367 Inplace ALTER on table with spatial idx might cause lost rows if concurrent purge
Description:
------------
A table with a secondary index on a spatial column may lose records
during an inplace ALTER TABLE operation if the records are deleted
and purged concurrently. This issue is specific to spatial indexes
and is similar to Bug#36846567, which was reported for tables with a
clustered index.

Analysis:
---------
During the inplace ALTER TABLE rebuild, the records in the clustered
index are scanned with the help of a cursor. Cursor tracks the record
position in the index. If the table has spatial index(es) then,
during the scan, its respective builder instance caches the records in
a tuple vector using the method Builder::batch_add_row().
batch_add_row() is called through a bulk_inserter lambda.
Once the cursor reaches at the end of the page, the cached tuples are
inserted (processed) into the spatial index(es) using the method
Builder::batch_insert(). batch_insert() is called through batch_inserter
lambda.
While processing the records, if the batch_inserter detects that the
server is running out of space in redo logs, it deep copies the
remaining user records in the tuple vector of current builder and from
the remaining other builders of spatial index(es) as well.
Batch_inserter than calls the savepoint() that first backs off the
cursor by one position to point to a user record, saves that new
position and, then releases the latch(es) held by the mini transaction.

When the scan resumes, cursor must be positioned to either, the user
record which comes next after the position saved before or, if there
is no such record, it must restore to supremum on the last leaf page.

At present PCursor::restore_from_savepoint() calls resume() to restore
the cursor position.

In the scenario reported, records are purged and the remaining records
are rearranged in the index. Here is what happens :

1. resume() calls PCursor::restore_position() which is wrapper over the
btr_pcur_t::restore_position().

2: btr_pcur_t::restore_position first restores the position on the
record saved before or its predecessor and returns false because, it is
not original record.

3: Then the caller of step#2, PCursor::restore_position() detects
the returned value false. It decides to move to the next record.

4: Finally, the caller of step#3, PCursor::resume() notices that cursor
is positioned on the last record of the page, then it moves the cursor
to the successor record.

Notice step#4 is assuming that step#3 restores the cursor to the saved
record(or its predecessor) but that is not case, as a result the record
where the cursor was positioned at step#3 is lost, how exactly?
Actually after step#3, the cursor is either
(a) positioned on the saved record, or if it was removed meanwhile:
(b) on a user record which comes after it or,
(c) a supremum record which comes after the hole which it left.
The logic in PCursor::resume() correctly handles case (a) by advancing,
case (c) by doing nothing, but fails in case of (b), as it advances the
cursor needlessly.

Fix:
----
As part of the Bug#36846567 fix, it was identified that the current
PCursor wrapper APIs attempted to handle multiple scenarios, making it
difficult to implement them correctly. The decision was made to separate
the APIs so that each performs a specific task neatly.

This fix introduces a similar set of save/restore APIs for the batch
inserter mode used for R-Tree index(es) scan:

- save_previous_user_record_as_last_processed:
  This save API stores the position of the last user record on the page.

- restore_to_first_unprocessed:
  This restore API positions the cursor to the record which comes next
  after the cursor position saved by the paired save API.

Change-Id: I89627e7904293dcc58ccbb8b8a64f73b56e8dc93