Bug#36575175: Hypergraph plan change causes "incorrect string value" ERROR
Some binlog tests using MyISAM started failing when running with the
hypergraph optimizer after WL#16107, which changed the plan chosen for
an INSERT SELECT statement so that the rows were inserted in a
different order. The insertion order matters because warnings raised
while inserting into non-transactional storage engines (such as
MyISAM) are upgraded to errors only if they are raised for the first
row. If inserting the first row succeeds, a warning raised for any
subsequent row is not upgraded to an error. The rationale is probably
that the already inserted rows cannot be rolled back (since the
storage engine is non-transactional), so the statement has had an
effect and should therefore not fail, as that could give the
impression that the statement didn't have any effect.

The failing statement tries to insert all possible two-byte sequences
into a column with cp932 encoding. Since not all two-byte sequences
are valid in the cp932 encoding, it is expected that warnings are
raised.

Fix the test by adding an ORDER BY clause to ensure that the rows are
inserted in the same order, regardless of which optimizer is used.
This ensures that the first byte sequence attempted inserted is
x'0000', which maps to a string consisting of two ASCII null
characters, so the insertion of the first row is successful. If, on
the other hand, the reverse ordering had been chosen, x'FFFF' would be
inserted first. This sequence is not valid in cp932, and it would
raise a warning that would be upgraded to an error.

Change-Id: Ib8385a6d1c1c666b8375b88669ec18b99862012d