WL#16117 Hypergraph cost model complete coverage for materialization, streaming, and aggregation
AccessPaths covered by this worklog are: TEMPTABLE_AGGREGATE,
MATERIALIZE, AGGREGATE, and STREAM.

For "Materialization without deduplication", always uses this cost
model.  For "Materialization with deduplication", use this cost model
for DISTINCT or GROUP BY, but for SET operations that use
deduplication, continue to use the old optimizer model. This is TODO
for future work.

Also use this cost model for the cost of table scan of the temporary
table used for materialization.

Do all the cost calculations using the formulae and cost constants
suggested by linear regression.  The linear regression demonstrated
that the execution time of all these paths increases with input rows,
number of aggregations, number of GROUP BY/DISTINCT fields, and also
with the output rows if deduplication is involved.

Support only in-memory temporary tables. This cost model needs to be
extended later for temporary tables on disk.

The machine used was a AMD EPYC 7J13 64-Core Processor with 2.6GHz
speed.

Added a JET test file for linear regression. This test suggests cost
estimation formula, and optionally verifies that the estimated:actual
cost ratios don't deviate too much,  and generates data for scatter
plots. The estimated cost is dependent on row count which is not
reliably accurate. So the test simulates the cost estimation by
calculating the cost using formulae suggested by linear regression and
using actual row count instead of estimated row count.

Change-Id: Iad938185327fa68d679499859b6eae08b10d1f3e