Bug#37272285 Assertion `false && "Inconsistent row counts for different AccessPath objects."' failed.
We calculate the number of result rows from an (inner) join as:

Rows(left_path) * Rows(right_path) * Selectivity(join_predicate)

If right_path is a ref access, then the join predicate may be partly
or fully evaluated through that ref access. For example, if we have:

SELECT * FROM left_tab JOIN right_tab ON
left_tab.field1=right_tab.field_with_index AND
left_tab.field2=right_tab.field_without_index

then Selectivity(join_predicate) should exclude the selectivity
"left_tab.field1=right_tab.field_with_index" and only contain the
selectivity of "left_tab.field2=right_tab.field_without_index" if the
former sub-predicate was mapped to a ref access.

The existing code covers the cases in which the sub-predicate that is
mapped to a ref access could also be used as a hash-join
condition. But there are also cases where the sub-predicate can be
used for ref access but not for hash join. In such cases, the
selectivity of the sub-predicate is counted twice, giving
inconsistent row counts and triggering the assert.

An example is:

SELECT * FROM t1 LEFT JOIN LATERAL
(SELECT ... FROM t2, t3 WHERE t3.field_with_index = t1.field1+t2.field2) d1
ON other_predicate

The predicate "t3.field_with_index = t1.field1+t2.field2" will not be
used as a hash-join condition between t2 and t3, as it depends on t1,
but it can be used in a ref access on t3 in "t1 NLJ (t2 NLJ t3)".

This commit ensures that we exclude the selectivity of a sub-predicate
that is mapped to a ref access, also when the sub-predicate cannot be
used as as hash-join condition.

Change-Id: I0b1c87556ac4af051d6a42ce4f3b6d5b83c412b7