BUG#33200087: DROP VIEW IF EXISTS results in error when               having a table with the same name
ANALYSIS:
=========
DROP VIEW IF EXISTS on a view which does not exist and which
has the same name as an existing table would result in an
error. Reporting an error in this scenario is restrictive
especially in a scenario where we are trying to drop
multiple views.

FIX:
====
A warning (ER_WRONG_OBJECT) is reported in such cases.

Note:
=====
This behavior was introduced in 8.0 by WL#6378: New data
dictionary (umbrella). As part of the patch, error handling
was made stricter for scenarios where the view did not
exist.

Change-Id: Ib0ee12b557d024ece5481537947db94d3585abb7