Bug#37350937 The routing_guidelines library broke macOS Xcode build
The patch for
    WL#14119 Routing Guidelines classifier
introduced an OBJECT library routing_guidelines-objects,
which is converted to a STATIC library routing_guidelines-static
and a SHARED library routing_guidelines.

This conversion does not work for macOS Xcode.
Use the OBJECT library everywhere, by introducing ALIASes.

Note that the OBJECT library contains output sources from Bison,
and this output can only be input to *one* target in Xcode.

Change-Id: I6f038860237ffc1876af8f9a1deb5154607031d6