Bug#37299071 ha_ndbcluster missing index handling escape
WL#14867 Open NDB table with missing index in MySQL added support
to the MySQL Ndb Cluster handler to open tables with missing
indexes.

This mechanism is intended to allow the features of the MySQL server
to be used to solve problems in cases where indexes cannot be rebuilt
due to unmet constraints.

With missing indexes, some of the SQL handler functionality is
unavailable - for example the use of indexes to efficiently select
rows [for modification], or to identify duplicates when processing
modifications, or to push joins relying on indexes etc.

In these situations it is acceptable for the server to return
an error.

This patch adds coverage of some missing cases to the
ndb_use_dropped_index testcase, and then modifies ha_ndbcluster
to handle these cases correctly.

Change-Id: If77c73982b1052ad0ad11bb4c82ed045c55997e4