Bug#37359213 AIO::release(Slot *slot) competes with AIO::reserve_slot slowing down parallel writes
The AIO (Asynchronous IO) object for writes maintains a fixed size
m_slots array. A page cleaner wishing to write a dirty page
back to tablespace has to reserve a slot and describe the async
operation in it. Once the operation is done, an io write completer
thread releases the slot. As slots are a limited resource and the
number of page cleaners can grow up to 64, while the number of io
completers (by default) is just 4, a situation may arise in which
io completers have to compete for AIO::m_mutex with much more
numerous page cleaners just to be able to release the slot they
are waiting for, a problem made worse by having to wake all of them
even though just one slot become available.

This patch solves this issue by introducing ut::counting_semaphore,
which is just like std::counting_semaphore with a difference that:
- acquire(foo) will call foo() whenever it sees the counter==0, which
  can be used to trigger os_aio_simulated_wake_handler_threads
- the counter is forced to be 32-bit to ensure efficient underlying
  implementation (in particular direct use of Futex on linux, which
  unlike Windows' WaitOnAddress supports only 32-bit memory
  locations)

It also contains a fallback implementation for platforms which do not
have std::atomic<T>::wait(T) (which also happen to not have
std::counting_semaphore neither).

The way it helps is that:
1. the fast-path is implemented using atomic operations such as
fetch_add and compare_exchange_weak which does not need mutual
exclusion in the fast-path case where we have enough slots and
wait isn't necessary.
2. In case wait is necessary it is implemented via std::atomic::wait()
and notify_one() which map directly to a Futex (or WaitOnAddress)
which perhaps internally takes some mutex, but still this might be
better than what InnoDB does now (first latch m_mutex on AIO, then
latch os_event's mutex, then call pthread_* method to sleep/wake up)
3. it wakes up just one thread, as opposed to waking up all of them.
In testing this turned out to be crucial difference.

Change-Id: Idfa875d89078338fba3793017e02993cc5e0e7f9