Bug#37275404 AT: testNodeRestart -n MultiCrashTest T1 fails occasionally in .4node4rpl
Context:
runMultiCrashTest crashes a subset of the running data nodes in
a cluster in a rolling restart fashion or in parallel, to check
that whether the situations where the cluster survives or the
cluster dies are as expected.

Problem:
In one of the cases (4 replica), test gracefully crashes 3 of 4
replicas in a node group and expects the 4th replica to die as
well (via a crash insertion in QMGR). Then test checks that all
nodes expected to be dead are dead and that the remaining nodes
are alive, finally test start all nodes again. Problem is that,
when nodes are started (via mgmapi) sometimes the 4th node is
not yet connected to the cluster and therefore the 'start'
command fails.

Solution:
Ensure that, before start the nodes via mgmapi, all the nodes
are already connected to the cluster.

Change-Id: Ib2d4265f4816bc6b975570f69aebfe3952e9bb96