Bug#37383098: Bad column definition in COM_STMT_EXECUTE Response
This is a regression from bug#36846564, from where an execute query
response for a prepared statement using cursors and having a BIT
column in the SELECT list contained an empty value for original
table name and original database name.

The fix is to add code in create_tmp_field_from_item() that copies
the correct values for these fields, the same way that it is done in
Field::new_field(). A more general solution would be appreciated, but
this is a minimal fix focused on solving the regression.

A test case in main.mysql_client_test has been added, but notice that
it does not cover the actual fix, since the C interface apparently
never uses the fields that are fixed here.

Change-Id: Iabf357f4e3a359293f9bab098c82b7023f1ad0de