Return the correct collation for temporal results
When functions like `GREATEST`, `LEAST` and `COALESCE` return a temporal
result, the collation is not set to `binary` as expected.

The collation for temporal types normally is always `binary` which is
the case for fields and for functions inheriting from
`Item_temporal_func`.

The comparison functions mentioned here though don't inherit from that
and can return various types. Internally they do set the numeric
collation correctly, but at the protocol level (and for what the
`COLLATION` function returns), it is translated to the binary collation.

This translation does not happen before this change for the given
functions, so we add it here.

Before:

```
mysql> select collation(date'2010-01-01');
+-----------------------------+
| collation(date'2010-01-01') |
+-----------------------------+
| binary                      |
+-----------------------------+
1 row in set (0.00 sec)

mysql> select collation(greatest(date'2010-01-01', date'2009-01-01'));
+---------------------------------------------------------+
| collation(greatest(date'2010-01-01', date'2009-01-01')) |
+---------------------------------------------------------+
| latin1_swedish_ci                                       |
+---------------------------------------------------------+
1 row in set (0.00 sec)
```

After:

```
mysql> select collation(date'2010-01-01');
+-----------------------------+
| collation(date'2010-01-01') |
+-----------------------------+
| binary                      |
+-----------------------------+
1 row in set (0.00 sec)

mysql> select collation(greatest(date'2010-01-01', date'2009-01-01'));
+---------------------------------------------------------+
| collation(greatest(date'2010-01-01', date'2009-01-01')) |
+---------------------------------------------------------+
| binary                                                  |
+---------------------------------------------------------+
1 row in set (0.00 sec)
```

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>